backend/apps/ledger/serializers.py
from rest_framework import serializers

from .models import LedgerEntry


class LedgerEntrySerializer(serializers.ModelSerializer):
    class Meta:
        model = LedgerEntry
        fields = (
            "id",
            "kind",
            "amount_paise",
            "payout",
            "description",
            "created_at",
        )
