backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.10
celery==5.4.0
redis==5.0.4
django-environ==0.11.2
gunicorn==22.0.0
whitenoise==6.6.0
dj-database-url==2.2.0
